// frontend/src/api/generated/models/bestEffortType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type BestEffortType = typeof BestEffortType[keyof typeof BestEffortType];


export const BestEffortType = {
  Distance1Km: 'Distance1Km',
  Distance5Km: 'Distance5Km',
  Distance10Km: 'Distance10Km',
  DistanceHalfMarathon: 'DistanceHalfMarathon',
  DistanceMarathon: 'DistanceMarathon',
  Power1Min: 'Power1Min',
  Power5Min: 'Power5Min',
  Power20Min: 'Power20Min',
} as const;